backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
motor==3.3.2
pymongo==4.6.1
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
httpx==0.27.0
bcrypt==4.1.3